+++ src/cdrizzleapi.c
@@ -23,27 +23,6 @@
 static PyObject *gl_Error;
 FILE *driz_log_handle = NULL;
 
-/** ---------------------------------------------------------------------------
- * Multiply each pixel in an image by a scale factor
- */
-
-static void
-scale_image(PyArrayObject *image, double scale_factor) {
-    long i, size;
-    float *imptr;
-
-    assert(image);
-    imptr = (float *)PyArray_DATA(image);
-
-    size = PyArray_DIMS(image)[0] * PyArray_DIMS(image)[1];
-
-    for (i = size; i > 0; --i) {
-        *imptr++ *= scale_factor;
-    }
-
-    return;
-}
-
 static int
 process_array_list(PyObject *list, integer_t *nx, integer_t *ny,
                    const char *name, PyArrayObject ***arrays, int *nmax,
@@ -192,15 +171,18 @@ static PyObject *
 tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
     (void)self;
 
-    const char *kwlist[] = {
-        "input",    "weights",  "pixmap",  "output",  "counts",   "context",
-        "input2",   "output2",  "dq",      "outdq",   "uniqid",   "xmin",
-        "xmax",     "ymin",     "ymax",    "scale",   "pixfrac",  "kernel",
-        "in_units", "expscale", "wtscale", "fillstr", "fillstr2", NULL};
+    const char *kwlist[] = {"input",        "weights",  "pixmap",  "output",
+                            "counts",       "context",  "input2",  "output2",
+                            "dq",           "outdq",    "uniqid",  "xmin",
+                            "xmax",         "ymin",     "ymax",    "iscale",
+                            "pscale_ratio", "scale",    "pixfrac", "kernel",
+                            "in_units",     "expscale", "wtscale", "fillstr",
+                            "fillstr2",     NULL};
 
     /* Arguments in the order they appear */
     PyObject *oimg, *owei, *pixmap, *oout, *owht, *ocon, *odq = NULL;
-    PyObject *oimg2 = NULL, *oout2 = NULL, *ooutdq = NULL;
+    PyObject *oimg2 = NULL, *oout2 = NULL, *ooutdq = NULL,
+             *opscale_ratio = NULL, *oscale = NULL;
     int i, n_none;
     int nsq_args, nsq_arr = 0, nsq_arr_out = 0;
 
@@ -209,7 +191,9 @@ tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
     integer_t xmax = 0;
     integer_t ymin = 0;
     integer_t ymax = 0;
-    double scale = 1.0;
+    float iscale = 1.0f;
+    float pscale_ratio = NPY_NANF;
+    float scale = NPY_NANF;
     double pfract = 1.0;
     char *kernel_str = "square";
     char *inun_str = "cps";
@@ -230,7 +214,6 @@ tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
     char *fillstr_end;
     bool_t do_fill, do_fill2;
     float fill_value, fill_value2;
-    float inv_exposure_time;
     struct driz_error_t error;
     struct driz_param_t p;
     integer_t size[2];
@@ -245,14 +228,61 @@ tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
     driz_error_init(&error);
 
     if (!PyArg_ParseTupleAndKeywords(
-            args, keywords, "OOOOOO|OOOOiiiiiddssffss:tdriz", (char **)kwlist,
+            args, keywords, "OOOOOO|OOOOiiiiifOOdssffss:tdriz", (char **)kwlist,
             &oimg, &owei, &pixmap, &oout, &owht, &ocon, /* OOOOOO */
             &oimg2, &oout2, &odq, &ooutdq,              /* OOOO */
             &uniqid, &xmin, &xmax, &ymin, &ymax,        /* iiiii */
-            &scale, &pfract, &kernel_str, &inun_str,    /* ddss */
+            &iscale, &opscale_ratio, &oscale,           /* fOO */
+            &pfract, &kernel_str, &inun_str,            /* dss */
             &expin, &wtscl, &fillstr, &fillstr2)        /* ffss */
     ) {
         return NULL;
+    }
+
+    if (oscale != NULL && oscale != Py_None) {
+        scale = (float)PyFloat_AsDouble(oscale);
+        if (PyErr_Occurred()) {
+            driz_error_set_message(&error, "Argument 'scale' is not a number.");
+            goto _exit;
+        }
+        if (scale <= 0.0f || !isfinite(scale)) {
+            driz_error_set_message(
+                &error, "Argument 'scale' must be positive and finite.");
+            goto _exit;
+        }
+        iscale = scale * scale;
+        pscale_ratio = scale;
+        if (py_warning(
+                PyExc_DeprecationWarning,
+                "Argument 'scale' has been deprecated since version 3.0 "
+                "and it will be removed in a future release. "
+                "Use 'iscale' and 'pscale_ratio' instead and set "
+                "iscale=pscale_ratio**2 to achieve the same effect as with "
+                "'scale'.") != 0) {
+            goto _exit;
+        }
+
+    } else {
+        if (iscale <= 0.0f || !isfinite(iscale)) {
+            driz_error_set_message(
+                &error, "Argument 'iscale' must be positive and finite.");
+            goto _exit;
+        }
+
+        if (opscale_ratio != NULL && opscale_ratio != Py_None) {
+            pscale_ratio = (float)PyFloat_AsDouble(opscale_ratio);
+            if (PyErr_Occurred()) {
+                driz_error_set_message(
+                    &error, "Argument 'pscale_ratio' is not a number.");
+                goto _exit;
+            }
+            if (pscale_ratio <= 0.0f || !isfinite(pscale_ratio)) {
+                driz_error_set_message(
+                    &error,
+                    "Argument 'pscale_ratio' must be positive and finite.");
+                goto _exit;
+            }
+        }
     }
 
     /* Get raw C-array data */
@@ -540,15 +570,7 @@ tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
 
     /* If the input image is not in CPS we need to divide by the exposure */
     if (inun != unit_cps) {
-        inv_exposure_time = 1.0f / expin;
-        scale_image(img, inv_exposure_time);
-        if (img2_list) {
-            for (i = 0; i < nsq_arr; ++i) {
-                if (img2_list[i] != NULL) {
-                    scale_image(img2_list[i], pow(inv_exposure_time, 2.0));
-                }
-            }
-        }
+        iscale /= expin;
     }
 
     /* Setup reasonable defaults for drizzling */
@@ -566,7 +588,8 @@ tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
     p.ymin = ymin;
     p.xmax = xmax;
     p.ymax = ymax;
-    p.scale = scale;
+    p.iscale = iscale;
+    p.pscale_ratio = pscale_ratio;
     p.pixel_fraction = pfract;
     p.kernel = kernel;
     p.in_units = inun;
@@ -594,9 +617,6 @@ tdriz(PyObject *self, PyObject *args, PyObject *keywords) {
         goto _exit;
     }
     if (driz_error_check(&error, "ymax must be > ymin", p.ymax > p.ymin)) {
-        goto _exit;
-    }
-    if (driz_error_check(&error, "scale must be > 0", p.scale > 0.0f)) {
         goto _exit;
     }
     if (driz_error_check(&error, "exposure time must be > 0",
@@ -661,22 +681,25 @@ static PyObject *
 tblot(PyObject *self, PyObject *args, PyObject *keywords) {
     (void)self;
 
-    const char *kwlist[] = {"source",  "pixmap", "output", "xmin",   "xmax",
-                            "ymin",    "ymax",   "scale",  "kscale", "interp",
-                            "exptime", "misval", "sinscl", NULL};
+    const char *kwlist[] = {"source",  "pixmap", "output", "xmin",
+                            "xmax",    "ymin",   "ymax",   "iscale",
+                            "kscale",  "scale",  "interp", "exptime",
+                            "fillval", "misval", "sinscl", NULL};
 
     /* Arguments in the order they appear */
     PyObject *oimg, *pixmap, *oout;
+    PyObject *oscale = NULL, *okscale = NULL, *oiscale = NULL;
+    PyObject *oef = NULL, *omisval = NULL, *ofillval = NULL;
     long xmin = 0;
     long xmax = 0;
     long ymin = 0;
     long ymax = 0;
-    double scale = 1.0;
-    float kscale = 1.0;
+    float scale = 1.0f;
+    float iscale = 1.0f;
     char *interp_str = "poly5";
-    float ef = 1.0;
-    float misval = 0.0;
-    float sinscl = 1.0;
+    float ef = 1.0f;
+    float fillval;
+    float sinscl = 1.0f;
 
     PyArrayObject *img = NULL, *out = NULL, *map = NULL;
     enum e_interp_t interp;
@@ -691,13 +714,117 @@ tblot(PyObject *self, PyObject *args, PyObject *keywords) {
     driz_error_init(&error);
 
     if (!PyArg_ParseTupleAndKeywords(
-            args, keywords, "OOO|lllldfsfff:tblot", (char **)kwlist, &oimg,
-            &pixmap, &oout,                    /* OOO */
-            &xmin, &xmax, &ymin, &ymax,        /* llll */
-            &scale, &kscale, &interp_str, &ef, /* dfsf */
-            &misval, &sinscl)                  /* ff */
+            args, keywords, "OOO|llllOOOsOOOf:tblot", (char **)kwlist, /* */
+            &oimg, &pixmap, &oout,                                     /* OOO */
+            &xmin, &xmax, &ymin, &ymax,                     /* llll */
+            &oiscale, &okscale, &oscale, &interp_str, &oef, /* fOOsO */
+            &ofillval, &omisval, &sinscl)                   /* OOf */
     ) {
         return NULL;
+    }
+
+    if (oscale != NULL && !Py_IsNone(oscale)) {
+        scale = (float)PyFloat_AsDouble(oscale);
+
+        if (PyErr_Occurred()) {
+            driz_error_set_message(&error, "Argument 'scale' is not a number.");
+            goto _exit;
+        }
+
+        if (scale <= 0.0f || !isfinite(scale)) {
+            driz_error_set_message(
+                &error, "Argument 'scale' must be positive and finite.");
+            goto _exit;
+        }
+        iscale = 1.0 / (scale * scale);
+
+        if (py_warning(PyExc_DeprecationWarning,
+                       "Argument 'scale' is deprecated, use 'iscale' "
+                       "instead and set it to 1.0 / (scale*scale).") != 0) {
+            goto _exit;
+        }
+
+    } else {
+        if (oiscale == NULL || Py_IsNone(oiscale)) {
+            iscale = 1.0f;
+        } else {
+            iscale = (float)PyFloat_AsDouble(oiscale);
+
+            if (PyErr_Occurred()) {
+                driz_error_set_message(&error,
+                                       "Argument 'iscale' is not a number.");
+                goto _exit;
+            }
+        }
+        if (iscale <= 0.0f || !isfinite(iscale)) {
+            driz_error_set_message(
+                &error, "Argument 'iscale' must be positive and finite.");
+            goto _exit;
+        }
+    }
+
+    if (okscale != NULL && !Py_IsNone(okscale)) {
+        if (py_warning(PyExc_DeprecationWarning,
+                       "Argument 'kscale' has been deprecated and it will be "
+                       "removed in a future version. It is no longer used by "
+                       "the blotting algorithm and can be safely ignored.") !=
+            0) {
+            goto _exit;
+        }
+    }
+
+    if (omisval != NULL && !Py_IsNone(omisval)) {
+        if (py_warning(PyExc_DeprecationWarning,
+                       "Argument 'misval' has been deprecated and has been "
+                       "replaced by 'fillval' to achieve the same effect.") !=
+            0) {
+            goto _exit;
+        }
+
+        fillval = (float)PyFloat_AsDouble(omisval);
+
+        if (ofillval != NULL && !Py_IsNone(ofillval)) {
+            driz_error_set_message(
+                &error,
+                "Argument 'fillval' should not be set when 'misval' is set.");
+            goto _exit;
+        }
+    } else {
+        if (ofillval != NULL && !Py_IsNone(ofillval)) {
+            fillval = (float)PyFloat_AsDouble(ofillval);
+
+            if (PyErr_Occurred()) {
+                driz_error_set_message(&error,
+                                       "Argument 'fillval' is not a number.");
+                goto _exit;
+            }
+        } else {
+            fillval = 0.0f;
+        }
+    }
+
+    if (oef != NULL && !Py_IsNone(oef)) {
+        if (py_warning(PyExc_DeprecationWarning,
+                       "Argument 'exptime' has been deprecated and it will be "
+                       "removed in a future version. Use 'iscale' to achieve "
+                       "the same.") != 0) {
+            goto _exit;
+        }
+
+        ef = (float)PyFloat_AsDouble(oef);
+
+        if (PyErr_Occurred()) {
+            driz_error_set_message(&error,
+                                   "Argument 'exptime' is not a number.");
+            goto _exit;
+        }
+        if (ef <= 0.0f || !isfinite(ef)) {
+            driz_error_set_message(
+                &error, "Argument 'exptime' must be positive and finite.");
+            goto _exit;
+        }
+    } else {
+        ef = 1.0f;
     }
 
     img = (PyArrayObject *)PyArray_ContiguousFromAny(oimg, NPY_FLOAT, 2, 2);
@@ -751,12 +878,11 @@ tblot(PyObject *self, PyObject *args, PyObject *keywords) {
     p.xmax = xmax;
     p.ymin = ymin;
     p.ymax = ymax;
-    p.scale = scale;
-    p.kscale = kscale;
+    p.iscale = iscale;
     p.in_units = unit_cps;
     p.interpolation = interp;
     p.ef = ef;
-    p.misval = misval;
+    p.fill_value = fillval;
     p.sinscl = sinscl;
     p.pixmap = map;
     p.error = &error;
@@ -773,12 +899,6 @@ tblot(PyObject *self, PyObject *args, PyObject *keywords) {
     if (driz_error_check(&error, "ymax must be > ymin", p.ymax > p.ymin)) {
         goto _exit;
     }
-    if (driz_error_check(&error, "scale must be > 0", p.scale > 0.0)) {
-        goto _exit;
-    }
-    if (driz_error_check(&error, "kscale must be > 0", p.kscale > 0.0)) {
-        goto _exit;
-    }
     if (driz_error_check(&error, "exposure time must be > 0", p.ef > 0.0)) {
         goto _exit;
     }
@@ -790,9 +910,9 @@ tblot(PyObject *self, PyObject *args, PyObject *keywords) {
 _exit:
     driz_log_message("ending tblot");
     driz_log_close(driz_log_handle);
-    Py_DECREF(img);
-    Py_DECREF(out);
-    Py_DECREF(map);
+    Py_XDECREF(img);
+    Py_XDECREF(out);
+    Py_XDECREF(map);
 
     if (driz_error_is_set(&error)) {
         if (strcmp(driz_error_get_message(&error), "<PYTHON>") != 0) {
@@ -994,12 +1114,13 @@ clip_polygon_wrap(PyObject *self, PyObject *args) {
 static struct PyMethodDef cdrizzle_methods[] = {
     {"tdriz", (PyCFunction)(void (*)(void))tdriz, METH_VARARGS | METH_KEYWORDS,
      "tdriz(image, weights, pixmap, output, counts, context, image2, "
-     "output2, dq, outdq, uniqid, xmin, xmax, ymin, ymax, scale, pixfrac, "
-     "kernel, in_units, expscale, wtscale, fillstr, fillstr2)"},
+     "output2, dq, outdq, uniqid, xmin, xmax, ymin, ymax, iscale, "
+     "pscale_ratio, pixfrac, kernel, in_units, expscale, wtscale, fillstr, "
+     "fillstr2)"},
     {"tblot", (PyCFunction)(void (*)(void))(PyCFunctionWithKeywords)tblot,
      METH_VARARGS | METH_KEYWORDS,
-     "tblot(image, pixmap, output, xmin, xmax, ymin, ymax, scale, kscale, "
-     "interp, exptime, misval, sinscl)"},
+     "tblot(image, pixmap, output, xmin, xmax, ymin, ymax, iscale, "
+     "interp, exptime, fillval, misval, sinscl)"},
     {"test_cdrizzle", (PyCFunction)test_cdrizzle, METH_VARARGS,
      "test_cdrizzle(data, weights, pixmap, output_data, output_counts)"},
     {"invert_pixmap", (PyCFunction)invert_pixmap_wrap, METH_VARARGS,